+++ examples/beep.rs
@@ -67,6 +67,8 @@ where
 
     //let c = pink();
 
+    //let c = fundsp::sound::pebbles();
+
     let c = c >> split::<U2>();
 
     //let c = fundsp::sound::risset_glissando(false);