+++ src/lib.rs
@@ -262,6 +262,7 @@ pub mod shape;
 pub mod signal;
 pub mod sound;
 pub mod svf;
+pub mod system;
 pub mod wave;
 pub mod wavetable;